Restructure the vite and vitest scripts
