+++ packages/beacon-node/vitest.config.ts
@@ -1,5 +1,5 @@
 import {defineConfig, mergeConfig} from "vitest/config";
-import {buildTargetPlugin} from "../../scripts/vitest/plugins/buildTargetPlugin.js";
+import {buildTargetPlugin} from "../../scripts/vite/plugins/buildTargetPlugin.js";
 import vitestConfig from "../../vitest.base.unit.config";
 
 export default mergeConfig(
