+++ scripts/vite/plugins/blsBrowserPlugin.ts
