+++ scripts/vite/plugins/buildTargetPlugin.ts
