+++ scripts/vite/polyfills/emptyModule.js
