+++ scripts/vite/polyfills/perf_hooks.js
