+++ scripts/vite/polyfills/persistent-merkle-tree.js
